remove `legacy` module
