--- randomizer/src/legacy/mod.rs
@@ -1,6 +1,0 @@
-/**
- * Files in this module are holdovers from the original ALBW Randomizer
- * TODO: Refactor to remove this entire module
- */
-pub mod graph;
-pub mod path;
